refactor(runtime): centralize agent error formatting
