+++ src/features/comments/error-format.test.ts
@@ -10,332 +10,76 @@ import {
   isLlmFetchError,
 } from './error-format.js'
 
-describe('comments/error-format', () => {
-  describe('formatErrorComment', () => {
-    it('formats rate limit error with warning icon and reset time', () => {
-      // #given a rate limit error
-      const error = createRateLimitError('API rate limit exceeded', new Date('2024-01-01T01:00:00Z'))
-
-      // #when formatting the error
-      const formatted = formatErrorComment(error)
-
-      // #then it should include warning icon
-      expect(formatted).toContain(':warning:')
-      // #then it should include the error type
-      expect(formatted).toContain('Rate Limit')
-      // #then it should include the message
-      expect(formatted).toContain('API rate limit exceeded')
-      // #then it should indicate it's retryable
-      expect(formatted).toContain('retryable')
-    })
-
-    it('formats LLM timeout error with clock icon', () => {
-      // #given an LLM timeout error
-      const error = createLLMTimeoutError('Model response timed out after 30s')
-
-      // #when formatting the error
-      const formatted = formatErrorComment(error)
-
-      // #then it should include timeout indicator
-      expect(formatted).toContain(':hourglass:')
-      // #then it should include the message
-      expect(formatted).toContain('Model response timed out')
-    })
-
-    it('formats fatal error with error icon', () => {
-      // #given an internal error (fatal)
-      const error = createErrorInfo('internal', 'Unexpected failure', false)
-
-      // #when formatting the error
-      const formatted = formatErrorComment(error)
-
-      // #then it should include error icon
-      expect(formatted).toContain(':x:')
-      // #then it should NOT indicate it's retryable
-      expect(formatted).not.toContain('retryable')
-    })
-
-    it('includes suggested action when provided', () => {
-      // #given an error with suggested action
-      const error = createErrorInfo('permission', 'Insufficient permissions', false, {
-        suggestedAction: 'Check repository permissions',
-      })
-
-      // #when formatting the error
-      const formatted = formatErrorComment(error)
-
-      // #then it should include the suggested action
-      expect(formatted).toContain('Check repository permissions')
-    })
-
-    it('includes details when provided', () => {
-      // #given an error with details
-      const error = createErrorInfo('api_error', 'API call failed', true, {
-        details: 'Status: 502 Bad Gateway',
-      })
-
-      // #when formatting the error
-      const formatted = formatErrorComment(error)
-
-      // #then it should include the details
-      expect(formatted).toContain('502 Bad Gateway')
-    })
+/**
+ * `src/features/comments/error-format.ts` is a thin re-export of the
+ * canonical `@fro-bot/runtime` error-format module (see
+ * `packages/runtime/src/agent/error-format/format.test.ts` for the full
+ * behavioral test suite). This file only proves the re-export surface
+ * still resolves and behaves identically to the deleted local
+ * implementation for existing callers.
+ */
+describe('comments/error-format (re-export smoke)', () => {
+  it('re-exports createRateLimitError producing a retryable rate_limit ErrorInfo', () => {
+    // #given a rate-limit message and a reset time
+    // #when building an ErrorInfo via the re-exported createRateLimitError
+    // #then it is classified as a retryable rate_limit error
+    const error = createRateLimitError('API rate limit exceeded', new Date('2024-01-01T01:00:00Z'))
+    expect(error.type).toBe('rate_limit')
+    expect(error.retryable).toBe(true)
   })
 
-  describe('createErrorInfo', () => {
-    it('creates error with required fields', () => {
-      // #when creating a basic error
-      const error = createErrorInfo('validation', 'Invalid input', false)
-
-      // #then it should have the correct structure
-      expect(error.type).toBe('validation')
-      expect(error.message).toBe('Invalid input')
-      expect(error.retryable).toBe(false)
-      expect(error.details).toBeUndefined()
-      expect(error.suggestedAction).toBeUndefined()
-    })
-
-    it('creates error with optional fields', () => {
-      // #when creating an error with all fields
-      const error = createErrorInfo('configuration', 'Missing config', false, {
-        details: 'auth.json not found',
-        suggestedAction: 'Add auth.json to secrets',
-      })
-
-      // #then it should include optional fields
-      expect(error.details).toBe('auth.json not found')
-      expect(error.suggestedAction).toBe('Add auth.json to secrets')
-    })
+  it('re-exports formatErrorComment producing markdown with icon and message', () => {
+    // #given an ErrorInfo built via the re-exported createLLMTimeoutError
+    // #when formatting it with the re-exported formatErrorComment
+    // #then the markdown includes the timeout icon and the original message
+    const error = createLLMTimeoutError('Model response timed out after 30s')
+    const formatted = formatErrorComment(error)
+    expect(formatted).toContain(':hourglass:')
+    expect(formatted).toContain('Model response timed out')
   })
 
-  describe('createRateLimitError', () => {
-    it('creates retryable rate limit error with reset time', () => {
-      // #given a reset time
-      const resetTime = new Date('2024-01-01T01:00:00Z')
-
-      // #when creating a rate limit error
-      const error = createRateLimitError('Rate limit exceeded', resetTime)
-
-      // #then it should be correctly structured
-      expect(error.type).toBe('rate_limit')
-      expect(error.retryable).toBe(true)
-      expect(error.resetTime).toEqual(resetTime)
-      expect(error.suggestedAction).toContain('wait')
-    })
+  it('re-exports createErrorInfo producing the requested error shape', () => {
+    // #given a type/message/retryable triple
+    // #when building an ErrorInfo via the re-exported createErrorInfo
+    // #then the fields are set as requested
+    const error = createErrorInfo('validation', 'Invalid input', false)
+    expect(error.type).toBe('validation')
+    expect(error.message).toBe('Invalid input')
+    expect(error.retryable).toBe(false)
   })
 
-  describe('createLLMTimeoutError', () => {
-    it('creates retryable LLM timeout error', () => {
-      // #when creating an LLM timeout error
-      const error = createLLMTimeoutError('Response timed out')
-
-      // #then it should be correctly structured
-      expect(error.type).toBe('llm_timeout')
-      expect(error.retryable).toBe(true)
-      expect(error.suggestedAction).toBeDefined()
-    })
+  it('re-exports createLLMFetchError producing a retryable llm_fetch_error', () => {
+    // #given a fetch failure message and model name
+    // #when building an ErrorInfo via the re-exported createLLMFetchError
+    // #then it is classified as a retryable llm_fetch_error with model details
+    const error = createLLMFetchError('fetch failed', 'opencode/big-pickle')
+    expect(error.type).toBe('llm_fetch_error')
+    expect(error.retryable).toBe(true)
+    expect(error.details).toContain('opencode/big-pickle')
   })
 
-  describe('createLLMFetchError', () => {
-    it('creates retryable LLM fetch error with model info', () => {
-      // #when creating an LLM fetch error
-      const error = createLLMFetchError('fetch failed', 'opencode/big-pickle')
-
-      // #then it should be correctly structured
-      expect(error.type).toBe('llm_fetch_error')
-      expect(error.retryable).toBe(true)
-      expect(error.message).toContain('fetch failed')
-      expect(error.details).toContain('opencode/big-pickle')
-      expect(error.suggestedAction).toBeDefined()
-    })
-
-    it('creates error without model when not provided', () => {
-      // #when creating an LLM fetch error without model
-      const error = createLLMFetchError('Connection refused')
-
-      // #then it should work without model details
-      expect(error.type).toBe('llm_fetch_error')
-      expect(error.message).toContain('Connection refused')
-      expect(error.details).toBeUndefined()
-    })
+  it('re-exports isLlmFetchError detecting network failure messages', () => {
+    // #given a fetch failure message
+    // #when checking it via the re-exported isLlmFetchError
+    // #then it is detected as an LLM fetch error
+    expect(isLlmFetchError('fetch failed')).toBe(true)
+    expect(isLlmFetchError('Invalid API key')).toBe(false)
   })
 
-  describe('isLlmFetchError', () => {
-    it('detects "fetch failed" error message', () => {
-      // #given various error messages
-      const errors = [
-        'fetch failed',
-        'FETCH FAILED',
-        'Request: fetch failed after 3 retries',
-        'Error: fetch failed due to network issue',
-      ]
-
-      // #then all should be detected as LLM fetch errors
-      for (const msg of errors) {
-        expect(isLlmFetchError(msg)).toBe(true)
-      }
-    })
-
-    it('detects connection timeout errors', () => {
-      // #given timeout-related error messages
-      const errors = ['ConnectTimeoutError', 'Connection timed out', 'connect timeout']
-
-      // #then all should be detected as LLM fetch errors
-      for (const msg of errors) {
-        expect(isLlmFetchError(msg)).toBe(true)
-      }
-    })
-
-    it('detects ECONNREFUSED and network errors', () => {
-      // #given network error messages
-      const errors = ['ECONNREFUSED', 'ECONNRESET', 'ETIMEDOUT', 'network error']
-
-      // #then all should be detected as LLM fetch errors
-      for (const msg of errors) {
-        expect(isLlmFetchError(msg)).toBe(true)
-      }
-    })
-
-    it('returns false for non-fetch errors', () => {
-      // #given non-network error messages
-      const errors = ['Invalid API key', 'Rate limit exceeded', 'Permission denied', 'Syntax error in prompt', '']
-
-      // #then none should be detected as LLM fetch errors
-      for (const msg of errors) {
-        expect(isLlmFetchError(msg)).toBe(false)
-      }
-    })
-
-    it('handles Error objects with message property', () => {
-      // #given an Error object
-      const error = new Error('fetch failed')
-
-      // #then it should detect the error
-      expect(isLlmFetchError(error)).toBe(true)
-    })
-
-    it('handles objects with cause property', () => {
-      // #given an error with cause
-      const error = {message: 'Request failed', cause: 'fetch failed'}
-
-      // #then it should detect via cause
-      expect(isLlmFetchError(error)).toBe(true)
-    })
-
-    it('handles null and undefined safely', () => {
-      // #then should return false without throwing
-      expect(isLlmFetchError(null)).toBe(false)
-      expect(isLlmFetchError(undefined)).toBe(false)
-    })
+  it('re-exports isAgentNotFoundError detecting agent-not-found messages', () => {
+    // #given an agent-not-found message
+    // #when checking it via the re-exported isAgentNotFoundError
+    // #then it is detected as an agent-not-found error
+    expect(isAgentNotFoundError('agent not found')).toBe(true)
+    expect(isAgentNotFoundError('fetch failed')).toBe(false)
   })
 
-  describe('formatErrorComment for LLM fetch error', () => {
-    it('formats LLM fetch error with warning icon', () => {
-      // #given an LLM fetch error
-      const error = createLLMFetchError('fetch failed', 'opencode/big-pickle')
-
-      // #when formatting the error
-      const formatted = formatErrorComment(error)
-
-      // #then it should include warning icon (retryable)
-      expect(formatted).toContain(':warning:')
-      // #then it should include the error type label
-      expect(formatted).toContain('LLM Fetch Error')
-      // #then it should indicate it's retryable
-      expect(formatted).toContain('retryable')
-    })
-  })
-
-  describe('isAgentNotFoundError', () => {
-    it('detects "agent not found" error messages', () => {
-      // #given various agent error messages
-      const errors = ['agent not found', 'Agent Not Found', 'Agent not found: sisyphus']
-
-      // #then all should be detected as agent errors
-      for (const msg of errors) {
-        expect(isAgentNotFoundError(msg)).toBe(true)
-      }
-    })
-
-    it('detects "unknown agent" error messages', () => {
-      // #given unknown agent messages
-      const errors = ['unknown agent: sisyphus', 'Unknown Agent']
-
-      // #then all should be detected as agent errors
-      for (const msg of errors) {
-        expect(isAgentNotFoundError(msg)).toBe(true)
-      }
-    })
-
-    it('detects "invalid agent" error messages', () => {
-      // #given invalid agent messages
-      const errors = ['invalid agent name', 'Invalid Agent']
-
-      // #then all should be detected as agent errors
-      for (const msg of errors) {
-        expect(isAgentNotFoundError(msg)).toBe(true)
-      }
-    })
-
-    it('detects "agent does not exist" patterns', () => {
-      // #given agent does not exist messages
-      const errors = ['agent sisyphus does not exist', 'Agent foo does not exist']
-
-      // #then all should be detected as agent errors
-      for (const msg of errors) {
-        expect(isAgentNotFoundError(msg)).toBe(true)
-      }
-    })
-
-    it('detects "no agent named" patterns', () => {
-      // #given no agent named messages
-      expect(isAgentNotFoundError('no agent named sisyphus')).toBe(true)
-    })
-
-    it('detects "agent is not available" patterns', () => {
-      // #given agent not available messages
-      expect(isAgentNotFoundError('agent sisyphus is not available')).toBe(true)
-    })
-
-    it('returns false for non-agent errors', () => {
-      // #given non-agent error messages
-      const errors = ['fetch failed', 'Rate limit exceeded', 'Permission denied', 'Invalid API key', '']
-
-      // #then none should be detected as agent errors
-      for (const msg of errors) {
-        expect(isAgentNotFoundError(msg)).toBe(false)
-      }
-    })
-
-    it('handles Error objects with message property', () => {
-      // #given an Error object
-      const error = new Error('agent not found')
-
-      // #then it should detect the error
-      expect(isAgentNotFoundError(error)).toBe(true)
-    })
-
-    it('handles objects with message property', () => {
-      // #given an object with message
-      const error = {message: 'unknown agent'}
-
-      // #then it should detect via message
-      expect(isAgentNotFoundError(error)).toBe(true)
-    })
-
-    it('handles null and undefined safely', () => {
-      // #then should return false without throwing
-      expect(isAgentNotFoundError(null)).toBe(false)
-      expect(isAgentNotFoundError(undefined)).toBe(false)
-    })
-  })
-
-  describe('createAgentError', () => {
-    it('creates non-retryable agent error with enable-omo suggestion', () => {
-      // #when creating an agent error
+  describe('createAgentError (local compatibility wrapper)', () => {
+    it('creates non-retryable agent error with enable-omo suggestion when agent is provided', () => {
+      // #when creating an agent error with an agent name
       const error = createAgentError('agent not found', 'sisyphus')
 
-      // #then it should be correctly structured
+      // #then it should be correctly structured with the historical oMo guidance
       expect(error.type).toBe('configuration')
       expect(error.retryable).toBe(false)
       expect(error.message).toContain('agent not found')
@@ -347,10 +91,11 @@ describe('comments/error-format', () => {
       // #when creating an agent error without agent name
       const error = createAgentError('unknown agent error')
 
-      // #then it should work without agent details
+      // #then it should work without agent details or oMo guidance
       expect(error.type).toBe('configuration')
       expect(error.message).toContain('unknown agent error')
       expect(error.details).toBeUndefined()
+      expect(error.suggestedAction).toBe('Verify the agent name is correct.')
       expect(error.suggestedAction).not.toContain('enable-omo')
     })
   })
